+++ R/mirai.R
@@ -22,10 +22,10 @@
     send_aio(ctx, data = as.raw(0L), mode = "serial")
     close(sock)
   })
-  envir <- recv_ctx(ctx, mode = "serial", keep.raw = FALSE)
+  envir <- recv(ctx, mode = "serial", keep.raw = FALSE)
   msg <- eval(expr = .subset2(envir, ".expr"), envir = envir)
   on.exit()
-  send_ctx(ctx, data = msg, mode = "serial", echo = FALSE)
+  send(ctx, data = msg, mode = "serial", echo = FALSE)
   Sys.sleep(2L)
   close(sock)
 
@@ -97,7 +97,8 @@ eval_mirai <- function(.expr, ...) {
     envir <- list2env(arglist)
     ctx <- context(daemons())
     aio <- request(ctx, data = envir, send_mode = "serial", recv_mode = "serial", keep.raw = FALSE)
-    .Call(mirai_create, aio, ctx)
+    `attr<-`(.subset2(aio, "aio"), "ctx", ctx)
+    `class<-`(aio, c("mirai", "recvAio"))
 
   } else {
 
@@ -114,8 +115,9 @@ eval_mirai <- function(.expr, ...) {
     sock <- socket(protocol = "req", listen = url, autostart = TRUE)
     ctx <- context(sock)
     aio <- request(ctx, data = envir, send_mode = "serial", recv_mode = "serial", keep.raw = FALSE)
-    attr(sock, "context") <- ctx
-    .Call(mirai_create, aio, sock)
+    `attr<-`(.subset2(aio, "aio"), "ctx", ctx)
+    `attr<-`(.subset2(aio, "aio"), "sock", sock)
+    `class<-`(aio, c("mirai", "recvAio"))
 
   }
 
@@ -210,17 +212,17 @@ call_mirai <- function(mirai) {
     close(sock)
     ..(.)
   })
-  while (TRUE) {
+  repeat {
     ctx <- context(sock)
-    envir <- recv_ctx(ctx, mode = "serial", keep.raw = FALSE)
+    envir <- recv(ctx, mode = "serial", keep.raw = FALSE)
     missing(envir) && break
     msg <- eval(expr = .subset2(envir, ".expr"), envir = envir)
-    send_ctx(ctx, data = msg, mode = "serial", echo = FALSE)
+    send(ctx, data = msg, mode = "serial", echo = FALSE)
     close(ctx)
   }
 
   on.exit()
-  send_aio(ctx, data = as.raw(1L), mode = "serial")
+  send_aio(ctx, data = as.raw(1L), mode = "raw")
   close(sock)
 
 }
@@ -346,7 +348,7 @@ daemons <- function(...) {
         res <- 0L
         for (i in seq_len(-delta)) {
           ctx <- context(sock)
-          aio <- request(ctx, data = .Call(mirai_scm), send_mode = "serial", recv_mode = "serial", keep.raw = FALSE)
+          aio <- request(ctx, data = .mirai_scm(), send_mode = "serial", recv_mode = "raw", keep.raw = FALSE)
           call_aio(aio)
           close(ctx)
           if (identical(.subset2(aio, "data"), as.raw(1L))) {