+++ tests/tests.R
@@ -1,10 +1,11 @@
 library(mirai)
+nanotest <- function(x) invisible(x || stop())
 
 m <- eval_mirai({
   q <- m + n + 1L
   q / m
 }, m = 2L, n = 3L)
-invisible(inherits(call_mirai(m), "mirai") || stop())
-invisible(m$data == 3L || stop())
-invisible(identical(call_mirai(m), m) || stop())
+nanotest(inherits(call_mirai(m), "mirai"))
+nanotest(m$data == 3L)
+nanotest(identical(call_mirai(m), m))